fix(tier10-float): Task 1 review fixes — overflow guard + house-style alignment
- Add GeometryOverflow variant + checked_mul guard against 32-bit
  stride×height wrap (wasm32, i686). Regression tests added.
- Add Eq + Hash derives to GbrFloatFrameError (matches house style).
- Make all 4 try_new constructors pub const fn with #[inline(always)].
- Change stride type from usize to u32 (matches GbrpFrame /
  GbrpHighBitFrame / Rgbf32Frame). Walker files updated with as usize
  casts at the indexing site.
- Replace explicit re-export list with pub use planar_gbr_float::* in
  src/frame/mod.rs.
- Use plain vec![] in frame tests (matches planar_gbr_high_bit.rs).
- Add padded-stride happy-path + height=1 edge-case tests.

Code-reviewer feedback resolved.

+++ src/yuv/gbrapf16.rs
@@ -79,10 +79,10 @@ pub fn gbrapf16_to<S: Gbrapf16Sink>(src: &Gbrapf16Frame<'_>, sink: &mut S) -> Re
   let b_plane = src.b();
   let r_plane = src.r();
   let a_plane = src.a();
-  let g_stride = src.g_stride();
-  let b_stride = src.b_stride();
-  let r_stride = src.r_stride();
-  let a_stride = src.a_stride();
+  let g_stride = src.g_stride() as usize;
+  let b_stride = src.b_stride() as usize;
+  let r_stride = src.r_stride() as usize;
+  let a_stride = src.a_stride() as usize;
 
   for row in 0..h {
     let g = &g_plane[row * g_stride..row * g_stride + w];
